backend/ml-pipeline/requirements.txt
pandas==2.1.3
scikit-learn==1.3.2
scipy>=1.11.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
plotly==5.18.0
reportlab==4.0.7
seabird==0.12.0
seaborn==0.13.0
pip-outdated==0.6.0
psycopg==3.1.14
psycopg-binary==3.1.14
sqlalchemy==2.0.23
werkzeug==3.0.1

# ===========================================================================
# OpenTelemetry — Distributed Tracing
# Auto-instruments Flask, psycopg, SQLAlchemy, and requests
# ===========================================================================
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-psycopg==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-requests==0.50b0